Consolidate multigroup code into main files
