+++ adult.py
@@ -2,32 +2,40 @@ import plot
 from data import *
 from plot import *
 from policies import *
-from pathlib import Path
+import warnings
+
+warnings.filterwarnings("error", category=RuntimeWarning)
 
 
-def main():
-    exp_prefix = "trial_"
-    group = "SEX"
-    density = 1  # 1 to load the full folktables dataset, smaller than one for a smaller portion.
-    n_samples_per_group = 500000  # smaller groups are discarded
-    T = 5000  # total number of rounds
-    n_seeds = 10  # number of runs of each policy (same data)
-    reg_param = 0.1  # regularization parameter for the ridge regression
-    noise_magnitude = None  # None to use the true rewards from the dataset as noisy
-    expl_coeff_oful = 0.1  # OFUL exploration coefficient. O is equivalent to Greedy.
-    alpha = 0.01  # Parameter for a fair policy that can be used as a baseline
+def main(
+    exp_prefix="trial",
+    group="RAC1P",
+    density=1,  # 1 to load the full folktables dataset, smaller than one for a smaller portion.
+    n_samples_per_group=5000,  # minimum number of samples for each group: smaller groups are discarded
+    T=2500,
+    n_arms=10,
+    n_seeds=1,  # number of runs of each policy (same data)
+    reg_param=0.01,  # regularization parameter for the ridge regression
+    noise_magnitude=0.2,  # None to use the true rewards from the dataset
+    expl_coeff_oful=0.01,  # OFUL exploration coefficient. O is equivalent to Greedy.
+    plot_mult=1.5,  # higher value gives a larger plot
+    plot_flag=False,
+    epsilon=0.1,
+    delta=1e-2,
+    L_tilde=1,
+    delta_tilde=1e-3,
+    noise_type="gaussian",
+):
 
-    plot_mult = 1.2  # higher value gives a larger plot
     exp_dir = (
-        f"exps/adult_new/{exp_prefix}g_{group}_d_{density}_n_{n_samples_per_group}_nm{noise_magnitude}"
-        f"_lambda_{reg_param}_T={T}_ns_{n_seeds}_ecOFUL{expl_coeff_oful}_alpha{alpha}/"
+        f"exps/adult/{exp_prefix}_na={n_arms}_T={T}_ns={n_seeds}_eps={epsilon}_del={delta}_Lt={L_tilde}_nt={noise_type}/"
     )
-
     run(
         exp_dir=exp_dir,
         plot_mult=plot_mult,
         group=group,
         density=density,
+        n_arms=n_arms,
         n_samples_per_group=n_samples_per_group,
         reg_param=reg_param,
         noise_magnitude=noise_magnitude,
@@ -35,34 +43,43 @@ def main():
         expl_coeff_oful=expl_coeff_oful,
         T=T,
         compute_density=True,
-        alpha=alpha,
+        plot_flag=plot_flag,
+        epsilon=epsilon,
+        delta=delta,
+        L_tilde=L_tilde,
+        noise_type=noise_type,
     )
-
-    plot.main(dir=f"{exp_dir}plots/", mult=plot_mult, mode_histogram="percentage")
+    if plot_flag:
+        plot.main(dir=f"{exp_dir}plots/", mult=plot_mult, mode_histogram="percentage")
 
 
 def run(
     # problem parameters,
     problem_seed=42,
-    group_bias=16.0,
     mu_noise_level=1e-8,
+    n_arms=10,
     compute_density=False,
-    group="SEX",
+    group="RAC1P",
     density=1,
-    n_samples_per_group=50000,
+    n_samples_per_group=5000,
     poly_degree=1,
     noise_magnitude=None,
-    alpha=0.1,
     # algo parameters
     reg_param=0.1,
-    expl_coeff_oful=1.0,
-    T=500,
-    # T = 5000,
+    expl_coeff_oful=0.1,
+    T=5000,  # total number of rounds
     algo_seeds=tuple(range(10)),
     plot_mult=0.8,
     exp_dir="exps/adult/simple/",
+    plot_flag=False,
+    epsilon=0.1,
+    delta=1e-2,
+    L_tilde=1,  # bound s.t. ||X||_2^2 + ||y||_2^2 <= L_tilde^2 -- normed to 1 in data.py
+    alpha_param=None,  # confidence parameter --- usual choice is 1/T according to Shariff & Sheffet 2018
+    delta_tilde=1e-3,
+    noise_type="gaussian",
 ):
-    assert T <= n_samples_per_group
+    from pathlib import Path
 
     Path(f"{exp_dir}plots/").mkdir(parents=True, exist_ok=True)
 
@@ -73,6 +90,7 @@ def run(
         n_samples_per_group=n_samples_per_group,
         seed=problem_seed,
         noise_magnitude=noise_magnitude,
+        n_arms=n_arms,
     )
     n_arms = P.n_arms
 
@@ -95,22 +113,32 @@ def run(
         save=True,
         dir=f"{exp_dir}plots/",
         suffix="adult_",
+        plot_flag=plot_flag,
     )
 
+    # confidence parameter --- usual choice is 1/T according to Shariff & Sheffet 2018
+    if alpha_param is None:
+        alpha_param = 1
     policies_generators = [
-        lambda: Random(),
+        # lambda: Random(),
         lambda: OFUL(reg_param, P.d, expl_coeff_oful),
+        lambda: PrivateFairGreedy(
+            T=T,
+            epsilon=epsilon,
+            delta=delta,
+            delta_tilde=delta_tilde,
+            L_tilde=L_tilde,
+            alpha_param=alpha_param,
+            noise_type=noise_type,
+            reg_param=reg_param,
+            d=P.d,
+            n_arms=n_arms,
+        ),
         lambda: FairGreedy(reg_param, P.d, mu_noise_level),
-        lambda: Greedy(reg_param, P.d),
+        # lambda: Greedy(reg_param, P.d),
         # lambda: FairGreedyKnownCDF(reg_param, P.d, mu_noise_level, P),
         # lambda: FairGreedyKnownMuStar(P)
         # lambda: FairGreedyNoNoise(reg_param, P.d),
-        # lambda: FairOFUL(reg_param, P.d, expl_coeff_oful),
-        # lambda: FairOFULKnownCDF(reg_param, P.d, P, expl_coeff_oful),
-        # lambda: FairLearnGreedy(
-        #     reg_param, P.d, r=np.ones(n_arms) / n_arms, alpha=T * alpha
-        # ),
-        # lambda: FairLearnOFUL(reg_param, P.d, r=np.ones(n_arms)/n_arms, alpha=T*alpha, expl_coeff=expl_coeff_oful),
     ]
 
     total_ps, total_dfs = [], []
@@ -129,19 +157,18 @@ def run(
         policy_name = dfs[0]["policy"].drop_duplicates()[0]
 
         print(f"Results for {policy_name}")
-        for df in dfs:
-            df.hist(
-                column="actions", bins=n_arms,
-            )
-            break
-
-        plt.title(f"action_histo_{policy_name}")
-        plt.show()
-        for m in ["pseudo_regret", "pseudo_fair_regret"]:
+        if plot_flag:
             for df in dfs:
-                df[m].plot()
-            plt.title(f"{m}_{policy_name}")
+                df.hist(column="sel_group", bins=P.n_groups)
+                break
+            plt.title(f"sel_group_histo_{policy_name}")
             plt.show()
+
+            for m in ["pseudo_regret", "pseudo_fair_regret"]:
+                for df in dfs:
+                    df[m].plot()
+                plt.title(f"{m}_{policy_name}")
+                plt.show()
         for p in ps:
             print(f"mu_est_MSE = {np.mean((p.get_mu_estimate() - P.mu_star) ** 2)}")
             print(f"mu_est = {p.get_mu_estimate()}")
@@ -158,10 +185,34 @@ def run(
         total_dfs.extend(dfs)
 
     result_df = pd.concat(total_dfs)
+
+    # Print start time for CSV writing
+    print(f"Writing {len(result_df)} rows to CSV using pandas...")
     result_df.to_csv(f"{exp_dir}results.csv")
+    print(f"results.csv saved to {exp_dir}")
 
     return exp_dir
 
 
+# NOTE: THIS IS WHERE EXPERIMENTS START
 if __name__ == "__main__":
-    main()
+    import argparse
+
+    parser = argparse.ArgumentParser()
+    parser.add_argument(
+        "--plot", action="store_true", help="Enable plotting (default: disabled)"
+    )
+    args = parser.parse_args()
+    for n_arms in (10,):
+        for n_samples_per_group in (5000,):
+            main(
+                n_arms=n_arms,
+                n_samples_per_group=n_samples_per_group,
+                n_seeds=1,
+                plot_flag=args.plot,
+                T=50000,
+                epsilon=15,
+                delta=0.1,
+                L_tilde=1,  # TODO: this needs to be max row norm of X+Y
+                delta_tilde=0.01,
+            )
